+++ src/components/combo_box/combo_box.test.js
@@ -9,6 +9,7 @@ describe('EuiComboBox', () => {
     const component = render(
       <EuiComboBox
         {...requiredProps}
+        searchValue=""
         onChange={() => {}}
         onSearchChange={() => {}}
         onCreateOption={() => {}}